+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 
 -   Get default functions.
+-   Restore single setting functions.
 
 ## [1.3.1] - 2023-01-30